backend/requirements.txt
crewai
fastapi
uvicorn
python-dotenv
langchain-openai
google-generativeai
duckduckgo-search
pytest
pytest-asyncio
httpx
langchain-community
